Move BootstrapPropertySource to its own class and expose delegate PropertySource via getter
